+++ lib/modules/cache/cache_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: cache
-// Updated at: 2026-05-06 20:49:23
-// Commit: feat(cache): add bi-directional event bus handler for cache
+// Updated at: 2026-05-09 09:29:28
+// Commit: refactor(cache): optimize data flow between domain and data layers
 
 class cacheComponentState {
-  final String commitRef = '958';
-  final String lastMessage = 'feat(cache): add bi-directional event bus handler for cache';
-  final DateTime timestamp = DateTime.parse('2026-05-06T20:49:23+05:00');
+  final String commitRef = '964';
+  final String lastMessage = 'refactor(cache): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-05-09T09:29:28+05:00');
 }